requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
playwright>=1.40.0
python-dotenv>=1.0.0
2captcha-python>=1.2.0
pydantic>=2.0.0
httpx>=0.25.0
